add a test to a class and delete the class
